Add build script.

+++ build.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
 
-# Path to local dev wordpress installation.
-wordpress_source=/Applications/XAMPP/htdocs/essycode
+# Load config. Should include:
+#   wordpress_source
+source config.cfg
 
 # Exit on error.
 set -e
@@ -10,26 +11,31 @@ set -e
 rm -rf build
 mkdir -p build
 
-# Copy Wordpress installation.
+# Copy Wordpress installation after checking that it exists.
+if [ ! -d assets/wordpress ]; then
+  echo ERROR: Could not find assets/wordpress. Run ./download-wordpress.sh and try again.
+  exit 1
+fi
 cp -r assets/wordpress/* build
+
+# Copy .htaccess from local installation.
+if [ -f $wordpress_source/.htaccess ]; then
+  cp $wordpress_source/.htaccess build/.htaccess
+fi
 
 # Replace sample configuration with live configuration.
 rm build/wp-config-sample.php
 cp assets/wp-config.php build
 
-# Remove default plugins and templates.
-rm -rf build/wp-content/plugins/akismet
-rm -f build/wp-content/plugins/hello.php
-rm -rf build/wp-content/themes/twentyfifteen
-rm -rf build/wp-content/themes/twentysixteen
-rm -rf build/wp-content/themes/twentyseventeen
-
-# Copy local dev plugins and themes.
+# Clear out default themes and plugins and copy from local dev.
+rm -rf build/wp-content/plugins/*
+rm -rf build/wp-content/themes/*
 cp -r $wordpress_source/wp-content/plugins/* build/wp-content/plugins
 cp -r $wordpress_source/wp-content/themes/* build/wp-content/themes
 
-# Copy elastic beanstalk extensions.
+# Copy elastic beanstalk extensions and remove sample configurations.
 cp -r assets/.ebextensions build
+rm -f build/.ebextensions/env-sample.config
 
 # Create artifact.
 cd build
